feat: authenticated records.
Data in the database is now signed with a fresh HMAC on every write, and authenticated by comparing HMACs on every read.

Also, an 'overview' of an item that excludes sensitive data is implemented so that not all the items are decrypted when the collection is fetched. Full decryption of an item will occur only when an item is opened.